+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.2] - 2018-??-??
 ### Added
 - `requirements.txt` file
+- helper function `issubset(list1, list2)` to improve readability
 
 ## [0.1.1] - 2018-08-28
 ### Added